# backend/swagger.yaml
openapi: 3.0.0
info:
  title: AURA API
  description: Autonomous AI agent backend for DeFi portfolio optimization on Avalanche
  version: 1.0.0
  contact:
    name: AURA Support
    email: api@aura-defi.xyz
    url: https://github.com/aura-defi/api
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3001/api/v1
    description: Development server
  - url: https://api.aura-defi.avalanche.network/api/v1
    description: Production server
...
paths:
...
  /backtesting/run:
    post:
      tags:
        - Backtesting
      summary: Run backtesting simulation
      description: Run historical simulation with specified parameters
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BacktestRequest'
      responses:
        '200':
          description: Backtesting completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BacktestResult'
  /backtesting/scenarios:
    post:
      tags:
        - Backtesting
      summary: Run scenario analysis
      description: Run multiple market scenario simulations
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                initialAmount:
                  type: number
                scenarios:
                  type: array
                  items:
                    type: string
                    enum: [bull, bear, normal]
      responses:
        '200':
          description: Scenario analysis completed
components:
  schemas:
...
    BacktestRequest:
      type: object
      required:
        - initialAmount
        - riskProfile
        - days
      properties:
        initialAmount:
          type: number
          minimum: 100
        riskProfile:
          type: string
          enum: [Conservative, Balanced, Aggressive]
        days:
          type: integer
          minimum: 1
          maximum: 365
        rebalanceFrequency:
          type: integer
          default: 14
        marketCondition:
          type: string
          enum: [bull, bear, normal]
          default: normal
    BacktestResult:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            finalValue:
              type: number
            totalReturn:
              type: number
            returnPercentage:
              type: number
            annualizedReturn:
              type: number
            sharpeRatio:
              type: number
            maxDrawdown:
              type: number
            volatility:
              type: number
            totalGasCosts:
              type: number
            rebalances:
              type: integer
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    
    WalletSignature:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from wallet signature authentication